backend/requirements.txt
# Backend dependencies
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-filter==24.3
django-cors-headers==4.4.0
psycopg[binary]==3.2.6
pytest==8.3.5
pytest-django==4.11.1
